+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/project/BeanTransformerConfiguration.java
@@ -1,7 +1,11 @@
 package com.dwarfeng.familyhelper.webapi.node.configuration.project;
 
 import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonTaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.JSFixedFastJsonProject;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.Project;
 import com.dwarfeng.familyhelper.project.stack.bean.entity.TaskTypeIndicator;
+import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.project.JSFixedFastJsonDispProject;
+import com.dwarfeng.familyhelper.webapi.stack.bean.disp.project.DispProject;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import org.dozer.Mapper;
@@ -15,6 +19,16 @@ public class BeanTransformerConfiguration {
 
     public BeanTransformerConfiguration(Mapper mapper) {
         this.mapper = mapper;
+    }
+
+    @Bean
+    public BeanTransformer<Project, JSFixedFastJsonProject> projectBeanTransformer() {
+        return new DozerBeanTransformer<>(Project.class, JSFixedFastJsonProject.class, mapper);
+    }
+
+    @Bean
+    public BeanTransformer<DispProject, JSFixedFastJsonDispProject> dispProjectBeanTransformer() {
+        return new DozerBeanTransformer<>(DispProject.class, JSFixedFastJsonDispProject.class, mapper);
     }
 
     @Bean